lttng-trace-clock-sh-code-rework
lttng trace clock sh code rework

LTTng trace clock definitions for SuperH were broken for the new Kernels.
The patch reviews this code using the clocksource API to handle trace clock.
For example, on ST Kernels the TMU channel 1 continues to be used as
clocksource and all works fine for LTT (tested on SH4 STM platforms based).
Concerning the TMU frequency, it comes from the module_clk clock available on
ST platforms.
This support was only tested on ST platforms and it is turned-on for
CPU_SH4 configurations. Other effort and tests are necessary to verify
it on other SH architectures.

From: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Giuseppe Cavallaro <peppe.cavallaro@st.com>
Signed-off-by: Mathieu Desnoyers <mathieu.desnoyers@efficios.com>