arm: user_regset: Crunch regs
Porting the patch to 2.6.37.6.

From: Roland McGrath <roland@redhat.com>
Date: Thu, 16 Apr 2009 19:03:48 -0700
Subject: [PATCH 13/17] arm: user_regset: Crunch regs

This converts the ptrace Crunch support into user_regset form.
There should be no change in the ptrace behavior on CONFIG_CRUNCH.

I am not set up to test any hardware with CONFIG_CRUNCH or to try
any software that uses Crunch so as to verify it.  I've tested that
it compiles.  This regset's code is as simple as any can be.

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Gabbasov <andrew_gabbasov@mentor.com>