arm: asm/syscall.h (unfinished)
Porting the patch to 2.6.37.6.

From: Roland McGrath <roland@redhat.com>
Date: Thu, 16 Apr 2009 21:57:13 -0700
Subject: [PATCH 16/17] arm: asm/syscall.h (unfinished)

This provides the user_stack_pointer() macro in asm/ptrace.h,
which some new arch-independent code wants to be able to use.

It also adds the asm/syscall.h header to let arch-independent
code interpret the registers as system call arguments or return.

The syscall_get_nr() function here is not really right.  I don't
know enough about ARM to finish it correctly.  It needs to figure
out if the blocked user task is really in the kernel for a system
call and return -1 if not.  I also did not try to handle all the
different ABI variants, which I don't really understand.

Until this is fixed, /proc/pid/syscall can show bogus results
for a process that is blocked inside a fault or something else
that's not a system call.  (It's probably all wrong for !AEABI too.)

Signed-off-by: Roland McGrath <roland@redhat.com>
Signed-off-by: Andrew Gabbasov <andrew_gabbasov@mentor.com>